+++ src/app/src/util/util.js
@@ -111,6 +111,7 @@ export const makeDashboardApiBlocksURL = () => '/api/api-blocks/';
 export const makeDashboardApiBlockURL = id => `/api/api-blocks/${id}/`;
 
 export const makeDashboardGeocoderURL = () => '/api/geocoder/';
+export const makeClaimGeocoderURL = id => `/api/facility-claims/${id}/geocode/`;
 
 export const makeDashboardActivityReportsURL = () =>
     '/api/facility-activity-reports/';